pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sqlcarbon"
version = "0.2.0"
description = "Reliable, deterministic SQL Server table-to-table copy tool"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["sql", "sqlserver", "etl", "data-migration", "database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]
dependencies = [
    "pyodbc>=4.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
    "pyarrow>=16.0",
]

[project.scripts]
sqlcarbon = "sqlcarbon.cli:cli"

[project.urls]
Homepage = "https://github.com/TroBeeOne/SQLCarbon"
Issues = "https://github.com/TroBeeOne/SQLCarbon/issues"

[tool.hatch.build.targets.wheel]
packages = ["sqlcarbon"]
